+++ Changelog.md
@@ -1,7 +1,7 @@
 ## Changelog
 ### 1.1.3.5-beta
 - **Fix** could not open GUI in multiplay
-- **Fix** Turbo Rotary Hearth Blast Smelter and Absolute Freezer did not be applied temperature bonus when `distinct bus` is true.
+- **Fix** Turbo Rotary Hearth Blast Smelter and Absolute Freezer were not applied temperature bonus when `distinct bus` is true.
 - **Change** the `distinct bus` in the absolute freezer has been disabled.
 ---